refactor: 백엔드 서버 모듈 구조 분리
- HTTP API 서버 단일 모듈을 config, routes, request, conversion, response, error 로 분리
- AppState 와 변환 동시성 제한기 구성을 별도 state 모듈로 정리
- 서버 모듈 구조 변경 내용을 architecture, testing, memory 문서에 반영
