+++ src/server/error.rs
@@ -1,0 +1,108 @@
+use axum::{
+    http::StatusCode,
+    response::{IntoResponse, Response},
+    Json,
+};
+use serde::Serialize;
+
+#[derive(Debug)]
+pub(super) struct ApiError {
+    pub(super) status: StatusCode,
+    pub(super) code: &'static str,
+    message: String,
+}
+
+#[derive(Serialize)]
+struct ErrorResponse<'a> {
+    error: ErrorDetail<'a>,
+}
+
+#[derive(Serialize)]
+struct ErrorDetail<'a> {
+    code: &'a str,
+    message: &'a str,
+}
+
+impl ApiError {
+    pub(super) fn bad_request(code: &'static str, message: impl Into<String>) -> Self {
+        Self {
+            status: StatusCode::BAD_REQUEST,
+            code,
+            message: message.into(),
+        }
+    }
+
+    pub(super) fn payload_too_large(message: impl Into<String>) -> Self {
+        Self {
+            status: StatusCode::PAYLOAD_TOO_LARGE,
+            code: "payload_too_large",
+            message: message.into(),
+        }
+    }
+
+    pub(super) fn timeout(message: impl Into<String>) -> Self {
+        Self {
+            status: StatusCode::REQUEST_TIMEOUT,
+            code: "conversion_timeout",
+            message: message.into(),
+        }
+    }
+
+    pub(super) fn too_many_requests(message: impl Into<String>) -> Self {
+        Self {
+            status: StatusCode::TOO_MANY_REQUESTS,
+            code: "conversion_queue_full",
+            message: message.into(),
+        }
+    }
+
+    pub(super) fn internal(message: impl Into<String>) -> Self {
+        Self {
+            status: StatusCode::INTERNAL_SERVER_ERROR,
+            code: "internal_error",
+            message: message.into(),
+        }
+    }
+
+    pub(super) fn image(error: image::ImageError) -> Self {
+        Self::bad_request(
+            "invalid_image",
+            format!("이미지를 읽을 수 없습니다: {error}"),
+        )
+    }
+
+    pub(super) fn conversion(error: crate::ConverterError) -> Self {
+        match error {
+            crate::ConverterError::Image(error) => Self::image(error),
+            crate::ConverterError::UnsupportedFormat(format) => Self::bad_request(
+                "unsupported_format",
+                format!("지원하지 않는 포맷입니다: {format}"),
+            ),
+            other => Self::internal(format!("이미지 변환에 실패했습니다: {other}")),
+        }
+    }
+}
+
+impl IntoResponse for ApiError {
+    fn into_response(self) -> Response {
+        let body = ErrorResponse {
+            error: ErrorDetail {
+                code: self.code,
+                message: &self.message,
+            },
+        };
+        (self.status, Json(body)).into_response()
+    }
+}
+
+impl From<std::io::Error> for ApiError {
+    fn from(error: std::io::Error) -> Self {
+        Self::internal(format!("입출력 오류: {error}"))
+    }
+}
+
+impl From<axum::extract::multipart::MultipartError> for ApiError {
+    fn from(error: axum::extract::multipart::MultipartError) -> Self {
+        Self::bad_request("invalid_multipart", format!("multipart 요청 오류: {error}"))
+    }
+}
